+++ src/main/java/io/github/codexrm/server/service/UserService.java
@@ -7,12 +7,8 @@ import io.github.codexrm.server.model.User;
 import io.github.codexrm.server.repository.RoleRepository;
 import io.github.codexrm.server.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
+import org.springframework.data.domain.*;
 import org.springframework.stereotype.Service;
-
 import jakarta.persistence.EntityNotFoundException;
 import java.util.HashSet;
 import java.util.List;
@@ -42,11 +38,13 @@ public class UserService {
     }
 
     public User get(Integer id) {
-        return userRepository.findById(id).get();
+        return userRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("User not found with id: " + id));
     }
 
     public String getPasswordById(Integer id) {
-        User user = userRepository.findById(id).get();
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("User not found with id: " + id));
         return user.getPassword();
     }
 
@@ -55,25 +53,28 @@ public class UserService {
     }
 
     public User update(User user) {
-        if (user.getId() != null && !userRepository.existsById(user.getId()))
-            throw new EntityNotFoundException("There is no entity with such ID in the database.");
-
+        if (user.getId() != null && !userRepository.existsById(user.getId())) {
+            throw new EntityNotFoundException("User not found with id: " + user.getId());
+        }
         return userRepository.save(user);
     }
 
     public void delete(Integer id) {
+        if (!userRepository.existsById(id)) {
+            throw new EntityNotFoundException("User not found with id: " + id);
+        }
         userRepository.deleteById(id);
     }
 
-    public String validateUser(String username, String email) {
-        String exist = "Valid";
-        if (userRepository.existsByUsername(username))
-            exist = "Error: Username is already taken!";
+    public void validateUser(String username, String email) {
 
-        if (userRepository.existsByEmail(email))
-            exist = "Error: Email is already in use!";
+        if (userRepository.existsByUsername(username)) {
+            throw new RuntimeException("Error: Username is already taken!");
+        }
 
-        return exist;
+        if (userRepository.existsByEmail(email)) {
+            throw new RuntimeException("Error: Email is already in use!");
+        }
     }
 
     public User createUserAccount(User user, boolean isUser, List<String> roleList) {
@@ -82,45 +83,41 @@ public class UserService {
 
         if (isUser) {
             Role userRole = roleRepository.findByName(ERole.ROLE_USER)
-                    .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
+                    .orElseThrow(() -> new EntityNotFoundException("Role ROLE_USER not found"));
             roles.add(userRole);
 
         } else {
             if (roleList == null) {
                 Role userRole = roleRepository.findByName(ERole.ROLE_USER)
-                        .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
+                        .orElseThrow(() -> new EntityNotFoundException("Role ROLE_USER not found"));
                 roles.add(userRole);
 
             } else {
                 roleList.forEach(role -> {
                     switch (role) {
                         case "ROLE_ADMIN":
-                            Role adminRole = roleRepository.findByName(ERole.ROLE_ADMIN)
-                                    .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
-                            roles.add(adminRole);
-
+                            roles.add(roleRepository.findByName(ERole.ROLE_ADMIN)
+                                    .orElseThrow(() -> new EntityNotFoundException("Role ROLE_ADMIN not found")));
                             break;
-                        case "ROLE_MANAGER":
-                            Role managerRole = roleRepository.findByName(ERole.ROLE_MANAGER)
-                                    .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
-                            roles.add(managerRole);
 
+                        case "ROLE_MANAGER":
+                            roles.add(roleRepository.findByName(ERole.ROLE_MANAGER)
+                                    .orElseThrow(() -> new EntityNotFoundException("Role ROLE_MANAGER not found")));
                             break;
 
                         case "ROLE_AUDITOR":
-                            Role auditorRole = roleRepository.findByName(ERole.ROLE_AUDITOR)
-                                    .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
-                            roles.add(auditorRole);
-
+                            roles.add(roleRepository.findByName(ERole.ROLE_AUDITOR)
+                                    .orElseThrow(() -> new EntityNotFoundException("Role ROLE_AUDITOR not found")));
                             break;
+
                         default:
-                            Role userRole = roleRepository.findByName(ERole.ROLE_USER)
-                                    .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
-                            roles.add(userRole);
+                            roles.add(roleRepository.findByName(ERole.ROLE_USER)
+                                    .orElseThrow(() -> new EntityNotFoundException("Role ROLE_USER not found")));
                     }
                 });
             }
         }
+
         user.setRoles(roles);
         return add(user);
     }
@@ -129,51 +126,23 @@ public class UserService {
 
         if (sort == null) {
             return new Sort.Order(Sort.Direction.ASC, "id");
-
-        } else {
-            switch (sort) {
-                case idAsc:
-                    return new Sort.Order(Sort.Direction.ASC, "id");
-
-                case idDesc:
-                    return new Sort.Order(Sort.Direction.DESC, "id");
-
-                case nameAsc:
-                    return new Sort.Order(Sort.Direction.ASC, "name");
-
-                case nameDesc:
-                    return new Sort.Order(Sort.Direction.DESC, "name");
-
-                case lastNameAsc:
-                    return new Sort.Order(Sort.Direction.ASC, "lastname");
-
-                case lastNameDesc:
-                    return new Sort.Order(Sort.Direction.DESC, "lastname");
-
-                case emailAsc:
-                    return new Sort.Order(Sort.Direction.ASC, "email");
-
-                case emailDesc:
-                    return new Sort.Order(Sort.Direction.DESC, "email");
-
-                case enabledAsc:
-                    return new Sort.Order(Sort.Direction.ASC, "enabled");
-
-                case enabledDesc:
-                    return new Sort.Order(Sort.Direction.DESC, "enabled");
-
-                case usernameAsc:
-                    return new Sort.Order(Sort.Direction.ASC, "username");
-
-                case usernameDesc:
-                    return new Sort.Order(Sort.Direction.DESC, "username");
-
-                case passwordAsc:
-                    return new Sort.Order(Sort.Direction.ASC, "password");
-
-                default:
-                    return new Sort.Order(Sort.Direction.DESC, "password");
-            }
         }
+
+        return switch (sort) {
+            case idAsc -> new Sort.Order(Sort.Direction.ASC, "id");
+            case idDesc -> new Sort.Order(Sort.Direction.DESC, "id");
+            case nameAsc -> new Sort.Order(Sort.Direction.ASC, "name");
+            case nameDesc -> new Sort.Order(Sort.Direction.DESC, "name");
+            case lastNameAsc -> new Sort.Order(Sort.Direction.ASC, "lastname");
+            case lastNameDesc -> new Sort.Order(Sort.Direction.DESC, "lastname");
+            case emailAsc -> new Sort.Order(Sort.Direction.ASC, "email");
+            case emailDesc -> new Sort.Order(Sort.Direction.DESC, "email");
+            case enabledAsc -> new Sort.Order(Sort.Direction.ASC, "enabled");
+            case enabledDesc -> new Sort.Order(Sort.Direction.DESC, "enabled");
+            case usernameAsc -> new Sort.Order(Sort.Direction.ASC, "username");
+            case usernameDesc -> new Sort.Order(Sort.Direction.DESC, "username");
+            case passwordAsc -> new Sort.Order(Sort.Direction.ASC, "password");
+            default -> new Sort.Order(Sort.Direction.DESC, "password");
+        };
     }
 }
